HL changes merged upto 25th July
git-svn-id: http://192.168.110.32/svn/pennapps/pff/root/branches/PLF-Release-5.10.0@146413 6963f9ea-81fa-44da-aa10-7ebde50aa541

+++ PFSJava/src/main/java/com/pennant/backend/dao/finance/covenant/impl/CovenantsDAOImpl.java
@@ -612,7 +612,7 @@ public class CovenantsDAOImpl extends SequenceDao<FinCovenantType> implements Co
 			c.setNextTaskId(rs.getString("NextTaskId"));
 			c.setRecordType(rs.getString("RecordType"));
 			c.setWorkflowId(rs.getLong("WorkflowId"));
-			c.setLastMntBy(rs.getInt("LastMntBy"));
+			c.setLastMntBy(rs.getLong("LastMntBy"));
 			c.setLastMntOn(rs.getTimestamp("LastMntOn"));
 
 			if (StringUtils.trimToEmpty(type).contains("View")) {
